+++ backend/src/scenarios/scenarios.test.ts
@@ -1,14 +1,12 @@
-import axios from 'axios';
 import { Server } from 'http';
 import express, { Express } from 'express';
 import { ScenarioAPIConfig, addScenarioApi } from './scenario.interface';
-import { Scenario, ScenarioDescription, ScenarioState } from './scenarios';
+import { ScenarioDescription, ScenarioState } from './scenarios';
 import { sleep } from '../utils/async';
 import { deleteFile } from '../utils/files';
 import { italyScenarioFactory } from '../usr/italy/italyScenario';
 
 
-const http = axios.create();
 const port = 5001;
 
 const config: ScenarioAPIConfig = {
@@ -43,30 +41,29 @@ afterAll(async () => {
 describe('scenarios', () => {
 
     test('GET scenarios', async () => {
-        const response = await http.get(`http://localhost:${port}/scenarios`);
-        const data = response.data;
+        const data = await (await fetch(`http://localhost:${port}/scenarios`)).json();
         expect(data).toBeTruthy();
         expect(data.length > 0).toBe(true);
         expect(data[0].id).toBeTruthy();
     });
 
     test('GET steps', async () => {
-        const response = await http.get(`http://localhost:${port}/scenarios`);
-        const scenarios = response.data;
+        const response = await fetch(`http://localhost:${port}/scenarios`);
+        const scenarios = await response.json();
         const scenario = scenarios[0];
-        const response2 = await http.get(`http://localhost:${port}/scenarios/${scenario.id}`);
-        const fullScenario = response2.data;
+        const response2 = await fetch(`http://localhost:${port}/scenarios/${scenario.id}`);
+        const fullScenario = await response2.json();
         expect(fullScenario).toBeTruthy();
         expect(fullScenario.steps.length > 0).toBe(true);
         expect(fullScenario.steps[0].title).toBeTruthy();
     });
 
     test('POST execute and poll', async () => {
-        const response = await http.get(`http://localhost:${port}/scenarios`);
-        const scenarios = response.data;
+        const response = await fetch(`http://localhost:${port}/scenarios`);
+        const scenarios = await response.json();
         const scenario = scenarios[0];
-        const response2 = await http.get(`http://localhost:${port}/scenarios/${scenario.id}`);
-        const fullScenario = response2.data;
+        const response2 = await fetch(`http://localhost:${port}/scenarios/${scenario.id}`);
+        const fullScenario = await response2.json();
         const step = fullScenario.steps[0];
         const inputs = step.inputs;
         const input = inputs[0];
@@ -77,16 +74,16 @@ describe('scenarios', () => {
             }]
         };
         // request 1: start execution
-        const response3 = await http.post(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute`, state);
-        const { ticket } = response3.data;
+        const response3 = await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute`, { body: JSON.stringify(state), method: 'POST' });
+        const { ticket } = await response3.json();
         expect(ticket).toBeTruthy();
         // request 2: first poll ...
-        const response4 = await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute/poll/${ticket}`);
-        expect(response4.data.ticket).toBeTruthy();
+        const response4 = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute/poll/${ticket}`)).json();
+        expect(response4.ticket).toBeTruthy();
         await sleep(1000);
         // request 3: second poll.
-        const response5 = await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute/poll/${ticket}`);
-        const { results: newState } = response5.data;
+        const response5 = await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute/poll/${ticket}`);
+        const { results: newState } = await response5.json();
         expect(newState).toBeTruthy();
         expect(newState.data.length > 0).toBe(true);
         expect(newState.data[0].id).toBeTruthy();
@@ -94,11 +91,11 @@ describe('scenarios', () => {
 
 
     test('POST 2nd step', async () => {
-        const response = await http.get(`http://localhost:${port}/scenarios`);
-        const scenarios = response.data;
+        const response = await fetch(`http://localhost:${port}/scenarios`);
+        const scenarios = await response.json();
         const scenario = scenarios[0];
-        const response2 = await http.get(`http://localhost:${port}/scenarios/${scenario.id}`);
-        const fullScenario = response2.data;
+        const response2 = await fetch(`http://localhost:${port}/scenarios/${scenario.id}`);
+        const fullScenario = await response2.json();
         const step = fullScenario.steps[0];
         const step2 = fullScenario.steps[1];
         const inputs = step.inputs;
@@ -110,17 +107,17 @@ describe('scenarios', () => {
                 value: input.options[1]
             }]
         };
-        const response3 = await http.post(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute`, state);
-        const { ticket } = response3.data;
+        const response3 = await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute`, { body: JSON.stringify(state), method: 'POST' });
+        const { ticket } = await response3.json();
         await sleep(1000);
-        const response4 = await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute/poll/${ticket}`);
-        const { results: newState } = response4.data;
-        const response5 = await http.post(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step2.id}/execute`, newState);
-        const { ticket: newTicket } = response5.data;
+        const response4 = await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute/poll/${ticket}`);
+        const { results: newState } = await response4.json();
+        const response5 = await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step2.id}/execute`, { body: JSON.stringify(newState), method: 'POST' });
+        const { ticket: newTicket } = await response5.json();
         expect(newTicket).toBeTruthy();
         await sleep(1000);
-        const response6 = await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step2.id}/execute/poll/${newTicket}`);
-        const { results: finalState } = response6.data;
+        const response6 = await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step2.id}/execute/poll/${newTicket}`);
+        const { results: finalState } = await response6.json();
         expect(finalState).toBeTruthy();
         expect(finalState.data.length > 1).toBe(true);
         expect(finalState.data[1].id).toBeTruthy();
@@ -137,7 +134,7 @@ describe('scenarios - cache', () => {
     beforeAll(async () => {
         await deleteFile(config.storeDir);
 
-        scenario = (await http.get(`http://localhost:${port}/scenarios/Italy`)).data;
+        scenario = await (await fetch(`http://localhost:${port}/scenarios/Italy`)).json();
         const step = scenario.steps[0];
         const inputs = step.inputs;
         const input = inputs[0];
@@ -149,10 +146,10 @@ describe('scenarios - cache', () => {
         };
 
         // request 1: start execution
-        const { ticket } = (await http.post(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute`, state)).data;
+        const { ticket } = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute`, { body: JSON.stringify(state), method: 'POST' })).json();
         await sleep(1000);
         // request 2: polling
-        const { results } = (await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute/poll/${ticket}`)).data;
+        const { results } = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step.id}/execute/poll/${ticket}`)).json();
         firstExecuteResult = results;
     })
 
@@ -160,9 +157,9 @@ describe('scenarios - cache', () => {
     test('cache works', async () => {
         // request 3: repeating same request
         const firstStep = scenario.steps[0];
-        const { ticket } = (await http.post(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute`, state)).data;
+        const { ticket } = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute`, { body: JSON.stringify(state), method: 'POST' })).json();
         await sleep(3);
-        const cachedResponse = (await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute/poll/${ticket}`)).data;
+        const cachedResponse = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute/poll/${ticket}`)).json();
         expect(cachedResponse.results).toBeTruthy();
         expect(cachedResponse.results).toEqual(firstExecuteResult);
     });
@@ -178,16 +175,16 @@ describe('scenarios - cache', () => {
             }]
         };
 
-        const { ticket } = (await http.post(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute`, newState)).data;
+        const { ticket } = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute`, { body: JSON.stringify(newState), method: 'POST' })).json();
         await sleep(3);
-        const response = (await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute/poll/${ticket}`)).data;
+        const response = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute/poll/${ticket}`)).json();
 
         // This was a new request, so there cannot be a result yet.
         expect(response.ticket).toBeTruthy();
         expect(response.results).toBeFalsy();
 
         await sleep(1000);
-        const { results } = (await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute/poll/${ticket}`)).data;
+        const { results } = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${firstStep.id}/execute/poll/${ticket}`)).json();
         expect(results).toBeTruthy();
         expect(results !== firstExecuteResult);
     });
@@ -212,7 +209,7 @@ describe('scenarios - in/out', () => {
 
     test('executing two jobs side-by-side.', async () => {
 
-        const scenario: ScenarioDescription = (await http.get(`http://localhost:${port}/scenarios/Italy`)).data;
+        const scenario: ScenarioDescription = await (await fetch(`http://localhost:${port}/scenarios/Italy`)).json();
         const step1 = scenario.steps.find(s => s.id === 'EqSim')!;
         const inputs = step1.inputs;
         const input = inputs[0];
@@ -223,11 +220,11 @@ describe('scenarios - in/out', () => {
             }]
         };
 
-        let response = (await http.post(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step1.id}/execute`, state)).data;
+        let response = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step1.id}/execute`, { body: JSON.stringify(state), method: 'POST' })).json();
         while (response.ticket) {
             await sleep(100);
             console.log("sleeping ...")
-            response = (await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step1.id}/execute/poll/${response.ticket}`)).data;
+            response = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step1.id}/execute/poll/${response.ticket}`)).json();
         }
         const newState = response.results;
         expect(newState).toBeTruthy();
@@ -235,21 +232,21 @@ describe('scenarios - in/out', () => {
         const step2 = scenario.steps.find(s => s.id === 'EqDmg')!;
         const step3 = scenario.steps.find(s => s.id === 'EqStats')!;
 
-        const result2 = await http.post(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step2.id}/execute`, newState);
-        const result3 = await http.post(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step3.id}/execute`, newState);
+        const result2 = await( await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step2.id}/execute`, { body: JSON.stringify(newState), method: 'POST' })).json();
+        const result3 = await( await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step3.id}/execute`, { body: JSON.stringify(newState), method: 'POST' })).json();
 
-        expect(result2.data.ticket).toBeTruthy();
-        expect(result3.data.ticket).toBeTruthy();
+        expect(result2.ticket).toBeTruthy();
+        expect(result3.ticket).toBeTruthy();
         
-        const ticket2 = result2.data.ticket;
-        const ticket3 = result3.data.ticket;
+        const ticket2 = result2.ticket;
+        const ticket3 = result3.ticket;
         expect(ticket2 !== ticket3).toBeTruthy();
 
 
         await sleep(500);
 
-        const response2 = (await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step2.id}/execute/poll/${ticket2}`)).data;
-        const response3 = (await http.get(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step3.id}/execute/poll/${ticket3}`)).data;
+        const response2 = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step2.id}/execute/poll/${ticket2}`)).json();
+        const response3 = await (await fetch(`http://localhost:${port}/scenarios/${scenario.id}/steps/${step3.id}/execute/poll/${ticket3}`)).json();
 
         expect(response2.results).toBeTruthy();
         expect(response3.results).toBeTruthy();