+++ backend/src/scenarios/scenarios.ts
@@ -126,6 +126,7 @@ export class Scenario {
         if (isResolvedDatum(entry)) return entry;
         if (isDatumReference(entry)) {
             const value = await this.store.getDataByKey(entry.reference);
+            if (value === undefined) throw Error(`Couldn't resolve datum "${id}" with refrence "${entry.reference}". Might be out of date.`);
             const datum: Datum = { id, value };
             return datum;
         }
@@ -200,6 +201,7 @@ export class Scenario {
 export class ScenarioFactory {
 
     private steps: Step[] = [];
+    private conditions: (() => Promise<true | string>)[] = [];
 
     constructor(public id: string, public description: string, public imageUrl?: string) {}
 
@@ -211,7 +213,20 @@ export class ScenarioFactory {
         this.steps.push(step);
     }
 
+    public registerCondition(condition: () => Promise<true | string>) {
+        this.conditions.push(condition);
+    }
+
+    public async verifyConditions() {
+        for(const condition of this.conditions) {
+            const fulfilled = await condition();
+            if (fulfilled !== true) return fulfilled;
+        }
+        return true;
+    }
+
     public createScenario(store: FileStorage<DatumLinage>) {
         return new Scenario(this.id, this.description, this.steps, store, this.imageUrl);
     }
+
 }