+++ compare-frontend/src/app/modules/map/converters/eqts/3_exposure.ts
@@ -4,7 +4,12 @@ import { Converter, LayerComposite } from "../../converter.service";
 import VectorLayer from "ol/layer/Vector";
 import VectorSource from "ol/source/Vector";
 import GeoJSON from 'ol/format/GeoJSON';
-import { StringPopupComponent } from "../../popups/string-popup/string-popup.component";
+import Style from "ol/style/Style";
+import Fill from "ol/style/Fill";
+import Stroke from "ol/style/Stroke";
+import { FeatureLike } from "ol/Feature";
+import { BarchartComponent } from "../../popups/barchart/barchart.component";
+import { BarData } from "src/app/helpers/d3charts";
 
 
 export class Exposure implements Converter {
@@ -17,20 +22,83 @@ export class Exposure implements Converter {
         const resolvedData = data.find(p => p.id === "exposure");
         if (!resolvedData || !resolvedData.value) return of([]);
 
+        const featureStyle = (feature: FeatureLike, resolution: number) => {
+
+            const props = feature.getProperties();
+            const expo = props['expo'];
+            let total = 0;
+            for (let i = 0; i < expo.Damage.length; i++) {
+                const nrBuildings = expo.Buildings[i];
+                total += nrBuildings;
+            }
+
+            let r = 160;
+            let g = 160;
+            let b = 160;
+            let a = 0.05;
+            if (total === 0) {
+                a = 0.9;
+            }
+      
+            return new Style({
+              fill: new Fill({
+                color: [r, g, b, a],
+      
+              }),
+              stroke: new Stroke({
+                color: [r, g, b, 1],
+                width: 2
+              })
+            });
+        }
+
+
         return of([{
             id: "exposureLayer",
-            visible: true,
+            opacity:  1.0,
             layer: new VectorLayer({
                 source: new VectorSource({
                     features: new GeoJSON({ dataProjection: 'EPSG:4326' }).readFeatures(resolvedData.value)
-                })
-            }),
-            popup: (location: number[]) => ({
-                component: StringPopupComponent,
-                args: {}  
+                }),
+                style: featureStyle
             }),
             onClick: () => {},
             onHover: () => {},
+            popup: (location, features) => {
+
+                if (features.length === 0) return undefined;
+
+                const props = features[0].getProperties();
+                const expo = props['expo'];
+
+                const data: BarData[] = [];
+                for (let i = 0; i < Object.values(expo.Taxonomy).length; i++) {
+                    const tax = expo['Taxonomy'][i]; // .match(/^[a-zA-Z]*/)[0];
+                    const bld = expo['Buildings'][i];
+                    if (!data.map(dp => dp.label).includes(tax)) {
+                        data.push({
+                        label: tax,
+                        value: bld,
+                        hoverText: `${bld} - {{ ${tax} }}`
+                        });
+                    } else {
+                        const datum = data.find(dp => dp.label === tax);
+                        if (datum) datum.value += bld;
+                    }
+                }
+
+                return {
+                    component: BarchartComponent,
+                    args: {
+                        data: data,
+                        width: 350,
+                        height: 300,
+                        xLabel: `Taxonomy`,
+                        yLabel: `Buildings`,
+                        title: `Exposure`
+                    }
+                }
+             }
         }]);
     }